requirements.txt
Pillow==4.3.0